# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Production API for Shiprion quotes, shipments, tracking, accounts, support, and operations.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /track/{trackingNumber}:
    get:
      operationId: trackShipment
      tags: [shipments]
      summary: Track a shipment by tracking number (public)
      description: Public endpoint — no authentication required.
      parameters:
        - name: trackingNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shipment and tracking events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentWithTracking"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /chat/sessions/{id}/messages:
    get:
      operationId: listChatMessages
      tags: [support]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          {
            description: Authorized chat messages,
            content:
              {
                application/json:
                  {
                    schema:
                      {
                        type: array,
                        items: { $ref: "#/components/schemas/ChatMessage" },
                      },
                  },
              },
          }
    post:
      operationId: createChatMessage
      tags: [support]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          {
            application/json:
              {
                schema:
                  {
                    type: object,
                    required: [content],
                    properties: { content: { type: string, maxLength: 2000 } },
                  },
              },
          }
      responses:
        "201":
          {
            description: Chat message created,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/ChatMessage" } },
              },
          }
  /chat/sessions/{id}/close:
    patch:
      operationId: closeChatSession
      tags: [support]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses: { "200": { description: Chat closed } }
  /chat/guest-sessions:
    post:
      operationId: createGuestChatSession
      tags: [support]
      requestBody:
        required: true
        content:
          {
            application/json:
              {
                schema:
                  {
                    type: object,
                    required: [guestName, guestEmail],
                    properties:
                      {
                        guestName:
                          { type: string, minLength: 2, maxLength: 100 },
                        guestEmail: { type: string, format: email },
                      },
                  },
              },
          }
      responses:
        "201":
          {
            description: Guest chat session and expiring bearer token,
            content:
              {
                application/json:
                  {
                    schema:
                      { $ref: "#/components/schemas/GuestChatSessionResponse" },
                  },
              },
          }
  /chat/guest-sessions/{id}/messages:
    get:
      operationId: listGuestChatMessages
      tags: [support]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - {
            name: X-Guest-Token,
            in: header,
            required: true,
            schema: { type: string },
          }
      responses:
        "200":
          {
            description: Guest chat messages,
            content:
              {
                application/json:
                  {
                    schema:
                      {
                        type: array,
                        items: { $ref: "#/components/schemas/ChatMessage" },
                      },
                  },
              },
          }
    post:
      operationId: createGuestChatMessage
      tags: [support]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - {
            name: X-Guest-Token,
            in: header,
            required: true,
            schema: { type: string },
          }
      requestBody:
        required: true
        content:
          {
            application/json:
              {
                schema:
                  {
                    type: object,
                    required: [content],
                    properties:
                      {
                        content:
                          { type: string, minLength: 1, maxLength: 2000 },
                      },
                  },
              },
          }
      responses:
        "201":
          {
            description: Guest message created,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/ChatMessage" } },
              },
          }
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
        requestId:
          type: string
      required: [code, message]
...
    ChatSession:
      type: object
      properties:
        id: { type: integer }
        userId: { type: [integer, "null"] }
        userEmail: { type: string }
        guestName: { type: [string, "null"] }
        status: { type: string, enum: [open, closed] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, userId, userEmail, guestName, status, createdAt, updatedAt]
    ChatMessage:
      type: object
      properties:
        id: { type: integer }
        sessionId: { type: integer }
        senderId: { type: [integer, "null"] }
        senderRole: { type: string }
        content: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, sessionId, senderId, senderRole, content, createdAt]
    GuestChatSessionResponse:
      type: object
      properties:
        session: { $ref: "#/components/schemas/ChatSession" }
        guestToken: { type: string }
        guestTokenExpiresAt: { type: string, format: date-time }
      required: [session, guestToken, guestTokenExpiresAt]
    PublicTrackingEvent:
      type: object
      description: >
        A tracking event as returned by the unauthenticated public endpoint.
        Coordinates (latitude/longitude) are intentionally omitted to prevent
        precise location intel from being accessible without authentication.
        Delivery-stage location strings are coarsened to destination city/country.
      properties:
        id:
          type: integer
        shipmentId:
          type: integer
        location:
          type: string
        status:
          type: string
        description:
          type: string
        occurredAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        [id, shipmentId, location, status, description, occurredAt, createdAt]
    ShipmentWithTracking:
      type: object
      description: >
        Public shipment tracking data. Intentionally omits recipient identity
        and precise location coordinates to protect recipient privacy.
      properties:
        id:
          type: integer
        trackingNumber:
          type: string
        origin:
          type: string
        destination:
          type: string
        status:
          type: string
          enum:
            [
              registered,
              pending,
              picked_up,
              in_transit,
              customs,
              on_hold_customs,
              arrived_at_port,
              out_for_delivery,
              delivered,
              cancelled,
            ]
        weightKg:
          type: ["number", "null"]
        estimatedDelivery:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        events:
          type: array
          items:
            $ref: "#/components/schemas/PublicTrackingEvent"
      required:
        - id
        - trackingNumber
        - origin
        - destination
        - status
        - weightKg
        - estimatedDelivery
        - createdAt
        - updatedAt
        - events
...